+++ tests/Translaas.Extensions.Mvc.Tests/TranslaasTagHelperTests.cs
@@ -220,7 +220,7 @@ public class TranslaasTagHelperTests
         // Arrange
         var mockService = new Mock<ITranslaasService>();
         mockService
-            .Setup(s => s.T(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int?>(), It.IsAny<CancellationToken>()))
+            .Setup(s => s.T(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<decimal?>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync("test");
 
         var tagHelper = new TranslaasTagHelper(mockService.Object);
